+++ lib/tables/scrollable_table.dart
@@ -21,6 +21,11 @@ class ScrollableTable<T> extends StatefulWidget {
     this.headerBackgroundColor,
     this.headerTextStyle,
     this.rowTextStyle,
+    this.customButtons = const [],
+    this.customButtonActions = const [],
+    this.buttonsWidth = 150,
+    this.minSizeButtons,
+    this.headerTextAlign,
   });
 
   final List<T> data;
@@ -39,6 +44,11 @@ class ScrollableTable<T> extends StatefulWidget {
   final Color? headerBackgroundColor;
   final TextStyle? headerTextStyle;
   final TextStyle? rowTextStyle;
+  final TextAlign? headerTextAlign;
+  final List<Widget> customButtons;
+  final List<void Function(T model)> customButtonActions;
+  final double buttonsWidth;
+  final double? minSizeButtons;
 
   @override
   State<ScrollableTable<T>> createState() => _ScrollableTableState<T>();
@@ -80,22 +90,15 @@ class _ScrollableTableState<T> extends State<ScrollableTable<T>> {
                         EdgeInsets.symmetric(horizontal: widget.columnSpacing),
                     child: Text(
                       widget.headers[i].label,
+                      textAlign: widget.headerTextAlign,
                       style: widget.headerTextStyle ??
                           PaginatedTableConfig.headerTextStyle,
                     ),
                   ),
                 ),
               if (widget.showButtons)
-                Container(
-                  width: 150,
-                  decoration: const BoxDecoration(
-                    border: Border(
-                      bottom: BorderSide(
-                        color: CupertinoColors.systemGrey,
-                        width: 1,
-                      ),
-                    ),
-                  ),
+                SizedBox(
+                  width: widget.buttonsWidth,
                   child: Center(
                     child: Text(
                       'Acciones',
@@ -150,7 +153,7 @@ class _ScrollableTableState<T> extends State<ScrollableTable<T>> {
                     ),
                     if (widget.showButtons)
                       SizedBox(
-                        width: 150,
+                        width: widget.buttonsWidth,
                         child: Row(
                           mainAxisAlignment: MainAxisAlignment.spaceEvenly,
                           children: [
@@ -162,6 +165,7 @@ class _ScrollableTableState<T> extends State<ScrollableTable<T>> {
                                       }
                                     : null,
                                 padding: EdgeInsets.zero,
+                                minSize: widget.minSizeButtons,
                                 child: const Icon(
                                   CupertinoIcons.pencil,
                                   color: CupertinoColors.activeBlue,
@@ -175,6 +179,7 @@ class _ScrollableTableState<T> extends State<ScrollableTable<T>> {
                                       }
                                     : null,
                                 padding: EdgeInsets.zero,
+                                minSize: widget.minSizeButtons,
                                 child: const Icon(
                                   CupertinoIcons.trash,
                                   color: CupertinoColors.destructiveRed,
@@ -188,10 +193,22 @@ class _ScrollableTableState<T> extends State<ScrollableTable<T>> {
                                       }
                                     : null,
                                 padding: EdgeInsets.zero,
+                                minSize: widget.minSizeButtons,
                                 child: const Icon(
                                   CupertinoIcons.info,
                                   color: CupertinoColors.activeGreen,
                                 ),
+                              ),
+                            for (int j = 0;
+                                j < widget.customButtons.length;
+                                j++)
+                              CupertinoButton(
+                                onPressed: () {
+                                  widget.customButtonActions[j](widget.data[i]);
+                                },
+                                padding: EdgeInsets.zero,
+                                minSize: widget.minSizeButtons,
+                                child: widget.customButtons[j],
                               ),
                           ],
                         ),