+++ CHANGELOG.md
@@ -1,3 +1,40 @@
+# 0.1.0
+* Upgrade deps
+
+## 0.0.9
+* Fix actions width in scrollable table
+
+## 0.0.8
+* Added headerTextAlign in tables
+
+## 0.0.7
+* Solved grey screen on map zoom exceeds limits
+
+## 0.0.6
+* Added anchorPos to markers
+
+## 0.0.5
+* Added minSizeButtons to tables
+
+## 0.0.4
+* Edit the zoom limit in widget creation
+* Send the Marker layer to front
+
+## 0.0.3
+
+* Limit the zoom to 21, to avoid grey screens
+* Added Satellital map provider
+
+## 0.0.2
+
+* Solved donut chart errors
+* Default grey chart on zero values on donut chart
+* Optimized donut chart corners 
+* Responsive donut chart text
+* Added MapController to map widget
+* Added CustomButtons to Tables
+
 ## 0.0.1
 
 * Initial release of the package.
+